Fix InsertTests
Summary:
Fixed failing InsertTime test in .NET connector.
- LastInsertedIdExplicitStart was deleted because feature 'CREATE TABLE...AUTO_INCREMENT=X for sharded tables' is not supported by S2
- LastInsertedIdInsertForeignKey was deleted because S2 doesn't support foreign keys (https://docs.singlestore.com/managed-service/en/developer-resources/unsupported-mysql-features/unsupported-feature-list.html)
- InsertMySqlSet: 'find_in_set' function was replaced with 'JSON_ARRAY_CONTAINS_STRING' because S2 doesn't have 'find_in_set' function
- InsertTime: previously the test used [InlineData(3)] instead of [InlineData(0)], but S2 doesn't support precision other than 0 or 6. When precision equals 0 S2 truncates TimeSpan value, as a result Assert.Equal() fails, that's why TimeSpan value was increased

Test Plan: https://app.circleci.com/pipelines/github/memsql/SingleStoreNETConnector/107/workflows/c2320cd0-2a9c-4e48-9aa1-7ccc11296cf6/jobs/113

Reviewers: pmishchenko-ua

Reviewed By: pmishchenko-ua

Subscribers: engineering-list

JIRA Issues: PLAT-6009

Differential Revision: https://grizzly.internal.memcompute.com/D54901

+++ tests/SideBySide/InsertTests.cs
@@ -154,37 +154,6 @@ UNLOCK TABLES;", m_database.Connection);
 		}
 	}
 
-	[SkippableFact(Baseline = "https://bugs.mysql.com/bug.php?id=97061")]
-	public async Task LastInsertedIdInsertForeignKey()
-	{
-		await m_database.Connection.ExecuteAsync(@"drop table if exists TestTableWithForeignKey;
-drop table if exists TestTable;
-
-Create Table TestTable(
-    id BIGINT NOT NULL AUTO_INCREMENT,
-    column1 CHAR(100),
-    Primary Key(id)
-);
-
-Create Table TestTableWithForeignKey(
-    foreign_id BIGINT NOT NULL,
-    column2 CHAR(100),
-    Foreign Key(foreign_id) REFERENCES TestTable(id)
-);");
-		try
-		{
-			await m_database.Connection.OpenAsync();
-			using var command = new SingleStoreCommand(@"INSERT INTO TestTable(column1) VALUES('hello');
-INSERT INTO TestTableWithForeignKey(foreign_id, column2) VALUES(LAST_INSERT_ID(), 'test');", m_database.Connection);
-			Assert.Equal(2, await command.ExecuteNonQueryAsync());
-			Assert.Equal(1L, command.LastInsertedId);
-		}
-		finally
-		{
-			m_database.Connection.Close();
-		}
-	}
-
 	[Fact]
 	public async Task RowsAffected()
 	{
@@ -199,24 +168,6 @@ INSERT INTO insert_rows_affected (value) VALUES (null);
 INSERT INTO insert_rows_affected (value) VALUES (null);", m_database.Connection);
 			var rowsAffected = await command.ExecuteNonQueryAsync();
 			Assert.Equal(2, rowsAffected);
-		}
-		finally
-		{
-			m_database.Connection.Close();
-		}
-	}
-
-	[Fact]
-	public void LastInsertedIdExplicitStart()
-	{
-		m_database.Connection.Execute(@"drop table if exists insert_ai_2;
-create table insert_ai_2(rowid integer not null primary key auto_increment, text varchar(100) not null) auto_increment = 1234;");
-		try
-		{
-			m_database.Connection.Open();
-			using var command = new SingleStoreCommand("insert into insert_ai_2(text) values('test');", m_database.Connection);
-			command.ExecuteNonQuery();
-			Assert.Equal(1234L, command.LastInsertedId);
 		}
 		finally
 		{
@@ -258,7 +209,10 @@ create table insert_time(value TIME({precision}));");
 			using (var reader = command.ExecuteReader())
 			{
 				Assert.True(reader.Read());
-				Assert.Equal(TimeSpan.FromMilliseconds(10), reader.GetValue(0));
+				if(precision == 0)
+					Assert.Equal(TimeSpan.Zero, reader.GetValue(0));
+				else
+					Assert.Equal(TimeSpan.FromMilliseconds(10), reader.GetValue(0));
 				Assert.False(reader.Read());
 			}
 		}
@@ -361,6 +315,11 @@ create table insert_stream(rowid integer not null primary key auto_increment, st
 	{
 		using var connection = new SingleStoreConnection(AppConfig.ConnectionString);
 		connection.Open();
+
+		Version utf8mb4_binSupportVersion = new(7, 5, 0);
+		if (connection.Session.S2ServerVersion.Version.CompareTo(utf8mb4_binSupportVersion) < 0)
+			return;
+
 		connection.Execute(@"drop table if exists insert_string_builder;
 create table insert_string_builder(rowid integer not null primary key auto_increment, str text collate utf8mb4_bin);");
 
@@ -674,10 +633,10 @@ create table insert_mysql_enums(
 		m_database.Connection.Execute(@"drop table if exists insert_mysql_set;
 create table insert_mysql_set(
 	rowid integer not null primary key auto_increment,
-	value set('one', 'two', 'four', 'eight') null
+	value set('""one""', '""two""', '""four""', '""eight""') null
 );");
-		m_database.Connection.Execute(@"insert into insert_mysql_set(value) values('one'), ('two'), ('one,two'), ('four'), ('four,one'), ('four,two'), ('four,two,one'), ('eight');");
-		Assert.Equal(new[] { "one", "one,two", "one,four", "one,two,four" }, m_database.Connection.Query<string>(@"select value from insert_mysql_set where find_in_set('one', value) order by rowid"));
+		m_database.Connection.Execute(@"insert into insert_mysql_set(value) values('""one""'), ('""two""'), ('""one"",""two""'), ('""four""'), ('""four"",""one""'), ('""four"",""two""'), ('""four"",""two"",""one""'), ('""eight""');");
+		Assert.Equal(new[] { "\"one\"", "\"one\",\"two\"", "\"one\",\"four\"", "\"one\",\"two\",\"four\"" }, m_database.Connection.Query<string>(@"select value from insert_mysql_set where JSON_ARRAY_CONTAINS_STRING(concat('[', value, ']'), 'one') order by rowid"));
 	}
